refactor(app): lift part/tool-call merging out of the streaming store
The 409 lines that turn a stream of deltas into a coherent `MessagePart[]` —
overlap-stripping on resume, post-tool text placement, tool-call
placeholders, revival, enriched-part merging, preview updates — move to
`v2-stream-parts.ts`. It is the most testable code in the file and none of
it is about being a store.

v2-streaming-store.ts 2,263 -> 1,871 lines.

This is deliberately a smaller cut than it could have been. A first attempt
took the whole 735-line helper region and had to be abandoned: `archiveCounter`
and `revisionFlushTarget` are module-level mutable state written from *both*
the helpers and the store body, so moving them would have meant exporting a
counter mutator and a target setter — making the module's internal contract
worse to buy a line count. That region is cohesive; this one genuinely is not.

What makes this seam real, and what the extraction script asserts:

  - no `let`/`var` crosses the boundary (the check that would have caught the
    first attempt automatically)
  - the region needed exactly two names from the rest of the file:
    `entryParts`, which came along with it, and the `AgentStreamEntry` type
  - the only edge back to the store is that type, imported `import type` and
    so erased at compile time

`finishUnresolvedTools` and `syncToolPartsForPersist` were part of the store's
public surface, so the store re-exports them and no import site changes.
